Move higher level functions load_filament_withSensor() load_filament_inPrinter() unload_filament_withSensor() to mmctl.cpp. Move led_blink() to main.cpp as it is not used anywhere else. Move isIdlerParked variable to motion.cpp and make it static, move its logic inside park_idler(), and home_idler() so it doesn't need to be accessed from the outside. Use get_idler_steps() and get_selector_steps() inside eject_filament() to simply its logic and remove duplicate code.
