+++ MM-control-01/mmctl.h
@@ -7,20 +7,17 @@
 
 extern bool isPrinting;
 extern int toolChanges;
-
 extern int active_extruder;
 extern int previous_extruder;
 extern bool isFilamentLoaded;
-extern bool isIdlerParked;
 
-
-extern void switch_extruder_withSensor(int new_extruder);
-
-extern void select_extruder(int new_extruder);
-
-extern bool feed_filament();
-
-void led_blink(int _no);
-
+void switch_extruder_withSensor(int new_extruder);
+void select_extruder(int new_extruder);
+bool feed_filament();
+void load_filament_withSensor();
+void load_filament_inPrinter();
+void unload_filament_withSensor();
+void eject_filament(int extruder);
+void recover_after_eject();
 
 #endif //_MMCTL_H
